refactor(qa): colocate runtime parity projection
