+++ extensions/qa-lab/src/suite-runtime-parity-runner.ts
@@ -11,11 +11,9 @@ import { sanitizeQaProgressValue as sanitizeQaSuiteProgressValue } from "./progr
 import type { QaThinkingLevel } from "./qa-gateway-config.js";
 import type { QaTransportAdapterFactory, QaTransportId } from "./qa-transport-registry.js";
 import {
-  isRuntimeParityResultPass,
   runRuntimeParityScenario,
   type RuntimeId,
   type RuntimeParityCell,
-  type RuntimeParityResult,
 } from "./runtime-parity.js";
 import { readQaBootstrapScenarioCatalog } from "./scenario-catalog.js";
 import type { QaScorecardChannelDriver, QaScorecardEvidenceMode } from "./scorecard-taxonomy.js";
@@ -26,6 +24,7 @@ import {
   resolveQaSuiteWorkerStartStaggerMs,
   scenarioRequiresControlUi,
 } from "./suite-planning.js";
+import { buildRuntimeParityScenarioResult } from "./suite-runtime-parity-result.js";
 import { remapModelRefForForcedRuntime } from "./suite-support.js";
 import type {
   QaSuiteRunParams,
@@ -39,88 +38,6 @@ import {
   requireQaSuiteStartLab,
   writeQaSuiteProgress,
 } from "./suite.js";
-
-function isRuntimeParityPass(result: RuntimeParityResult) {
-  return isRuntimeParityResultPass(result);
-}
-
-function formatRuntimeParityCellDetails(cell: RuntimeParityCell) {
-  const errors = [cell.transportErrorClass, cell.runtimeErrorClass].filter(Boolean).join(", ");
-  const sentinels = cell.sentinelFindings?.map((finding) => finding.kind).join(", ");
-  return [
-    `runtime=${cell.runtime}`,
-    `wallMs=${cell.wallClockMs}`,
-    `toolCalls=${cell.toolCalls.length}`,
-    `finalChars=${cell.finalText.length}`,
-    `tokens=${cell.usage.totalTokens}`,
-    ...(errors ? [`errors=${errors}`] : []),
-    ...(sentinels ? [`sentinels=${sentinels}`] : []),
-  ].join(" ");
-}
-
-function formatRuntimeParityScenarioCellDetails(cell: RuntimeParityResult["cells"][RuntimeId]) {
-  return [cell.details, formatRuntimeParityCellDetails(cell)].filter(Boolean).join("\n");
-}
-
-function runtimeParityScenarioStepStatus(
-  cell: Pick<
-    RuntimeParityResult["cells"][RuntimeId],
-    "runtimeErrorClass" | "status" | "transportErrorClass"
-  >,
-) {
-  if (cell.status === "fail" || cell.runtimeErrorClass || cell.transportErrorClass) {
-    return "fail";
-  }
-  if (cell.status === "skip") {
-    return "skip";
-  }
-  return "pass";
-}
-
-function runtimeParityScenarioResultStatus(result: RuntimeParityResult) {
-  const cellStatuses = new Set([
-    runtimeParityScenarioStepStatus(result.cells.openclaw),
-    runtimeParityScenarioStepStatus(result.cells.codex),
-  ]);
-  if (cellStatuses.has("fail")) {
-    return "fail";
-  }
-  if (cellStatuses.has("skip")) {
-    return "skip";
-  }
-  return isRuntimeParityPass(result) ? "pass" : "fail";
-}
-
-function buildRuntimeParityScenarioResult(params: {
-  scenarioName: string;
-  result: RuntimeParityResult;
-}): QaSuiteScenarioResult {
-  const driftStepStatus = runtimeParityScenarioResultStatus(params.result);
-  const openclawCell = params.result.cells.openclaw;
-  return {
-    name: params.scenarioName,
-    status: driftStepStatus,
-    details: params.result.driftDetails ?? `runtime drift classified as ${params.result.drift}`,
-    steps: [
-      {
-        name: openclawCell.runtime,
-        status: runtimeParityScenarioStepStatus(openclawCell),
-        details: formatRuntimeParityScenarioCellDetails(openclawCell),
-      },
-      {
-        name: params.result.cells.codex.runtime,
-        status: runtimeParityScenarioStepStatus(params.result.cells.codex),
-        details: formatRuntimeParityScenarioCellDetails(params.result.cells.codex),
-      },
-      {
-        name: "runtime drift",
-        status: driftStepStatus,
-        details: params.result.driftDetails ?? params.result.drift,
-      },
-    ],
-    runtimeParity: params.result,
-  };
-}
 
 export async function runQaRuntimeParitySuite(params: {
   runQaFlowSuite: QaSuiteRunner;
@@ -372,9 +289,3 @@ export async function runQaRuntimeParitySuite(params: {
     }
   }
 }
-
-export const qaSuiteRuntimeParityTesting = {
-  formatRuntimeParityScenarioCellDetails,
-  runtimeParityScenarioResultStatus,
-  runtimeParityScenarioStepStatus,
-};
